requirements.txt
Flask==1.1.2
gunicorn==20.1.0
eventlet==0.30.2

psycopg2==2.8.6
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.7.0

Flask-Mailman==0.2.2
